Major changes
Removed playerchallenger, it was a mess and unnecessarily object-oriented. Started tracing.py to replace it with soemthing more readable. tracing.py is not yet functional and has no unit tests. graphtheory and simulator have both minor changes and are still passing unit tests. ring member selection is still a PITA.